refactor(react-query-devtools): move ActiveQuery to subdirectory
